requirements.txt
openai
requests
praw
python-dotenv
Flask
yfinance